refactoring PostCollectionViewCell
