+++ GuidancerUI/CollectionViewCell/PostCollectionViewCell.swift
@@ -7,10 +7,10 @@
 
 import UIKit
 
-class PostCollectionViewCell: UICollectionViewCell {
+class PostCollectionViewCell: GCollectionViewCell {
     
-     let informationInPost = DescriptionPostSV()
-     let descriptionOfPost = ChangableInfoOfPostSV()
+    private let informationInPost = DescriptionPostSV()
+    let contentStackView = UIStackView()
     
     override init(frame: CGRect) {
         super.init(frame: frame)
@@ -22,13 +22,13 @@ class PostCollectionViewCell: UICollectionViewCell {
         fatalError("init(coder:) has not been implemented")
     }
     
-     func setup() {
-         let sv = UIStackView(arrangedSubviews: [informationInPost, descriptionOfPost])
-        sv.axis = .vertical
-        sv.embed(in: contentView)
+    func setup() {
+        contentStackView.addArrangedSubview(informationInPost)
+        contentStackView.axis = .vertical
+        contentStackView.embed(in: contentView)
     }
     
-     func cornerAndShadow() {
+    private  func cornerAndShadow() {
         self.layer.cornerRadius = 25
         self.backgroundColor = .white
         self.layer.shadowOffset = CGSize(width: 2.5, height: 2.5)
@@ -37,47 +37,18 @@ class PostCollectionViewCell: UICollectionViewCell {
         self.layer.shadowColor = UIColor.gBlack?.cgColor
     }
     
-    func setupContentForSearch(image: UIImage?,
+    func setupContent(image: UIImage?,
                       name: String,
-                      descriptionOfPlace: String,
-                      whoAndWhenPost: String,
-                      profileUserImageView: UIImage?,
-                      likes: Int,
-                      comments: Int
-    ) {
-        informationInPost.setupContent(image: image,
-                                       bestPlaceOfCityName: name,
-                                       descriptionOfPlace: descriptionOfPlace)
-        descriptionOfPost.setupContent(whoAndWhenPost: whoAndWhenPost,
-                                       profileUserImageView: profileUserImageView,
-                                       likes: likes,
-                                       comments: comments)
-    }
-    
-    func setupContentForProfile(image: UIImage?,
-                                 name: String,
-                                 descriptionOfPlace: String
+                      descriptionOfPlace: String
     ) {
         informationInPost.setupContent(image: image,
                                        bestPlaceOfCityName: name,
                                        descriptionOfPlace: descriptionOfPlace)
     }
     
-    class DescriptionPostSV: UIStackView {
-        
-//        private let viewMoreButton: GButton = {
-//            let bt = GButton { _ in
-//
-//            }
-//            bt.setTitleColor(.orange, for: .normal)
-//            bt.setTitle("View more...", for: .normal)
-//            bt.titleLabel?.font = .systemFont(ofSize: 13)
-//            return bt
-//        }()
-                             
-                             
-                             
-         var imageView: UIImageView = {
+    private final class DescriptionPostSV: UIStackView {
+
+        var imageView: UIImageView = {
             let image = UIImageView()
             image.clipsToBounds = true
             image.layer.cornerRadius = 25.0
@@ -86,13 +57,9 @@ class PostCollectionViewCell: UICollectionViewCell {
             return image
         }()
         
-         let bestPlaceOfCityName = GLabel(text: "Лучшие итальянские рестораны Санкт-Петербурга",
-                                                 font: UIFont.systemFont(ofSize: 17, weight: .regular),
-                                                 numberOfLines: 0)
+        let bestPlaceOfCityName = GLabel(font: UIFont.systemFont(ofSize: 17, weight: .regular))
         
-         let descriptionOfPlace = GLabel(text: "В этой прогулке я расскажу вам о лучших итальяннских ресторанах города СПб. Да, это не аудиопрогулка, но зато какой интересный рассказ...",
-                                                font: UIFont.systemFont(ofSize: 15, weight: .light),
-                                                numberOfLines: 0)
+        let descriptionOfPlace = GLabel(font: UIFont.systemFont(ofSize: 15, weight: .light))
         
         override init(frame: CGRect) {
             super.init(frame: frame)
@@ -102,18 +69,8 @@ class PostCollectionViewCell: UICollectionViewCell {
         required init(coder: NSCoder) {
             fatalError("init(coder:) has not been implemented")
         }
-        
-  
-//        private func viewMore() {
-//            let sizeOfTextFrame = descriptionOfPlace.frame.size.height
-//            if sizeOfTextFrame > sizeOfTextFrame + 1 {
-//
-//            }
-//
-//        }
-     
-         func setup() {
-   
+
+        private func setup() {
             spacing = 5
             distribution = .equalCentering
             axis = .vertical
@@ -121,7 +78,6 @@ class PostCollectionViewCell: UICollectionViewCell {
             addArrangedSubview(imageView)
             addArrangedSubview(bestPlaceOfCityName)
             addArrangedSubview(descriptionOfPlace)
-//            addArrangedSubview(viewMoreButton)
             NSLayoutConstraint.activate([
                 descriptionOfPlace.leadingAnchor.constraint(equalTo: leadingAnchor, constant: 10),
                 bestPlaceOfCityName.leadingAnchor.constraint(equalTo: leadingAnchor, constant: 10)
@@ -138,65 +94,6 @@ class PostCollectionViewCell: UICollectionViewCell {
         }
     }
     
-    
-    class ChangableInfoOfPostSV: UIView {
-        
-         var profileUserImageView: UIImageView = {
-            let image = UIImageView()
-            image.frame.size = CGSize(width: 50, height: 50)
-             image.clipsToBounds = true
-            return image
-        }()
-        override func layoutSubviews() {
-            super.layoutSubviews()
-            NSLayoutConstraint.activate([
-                profileUserImageView.heightAnchor.constraint(equalToConstant: 50),
-                profileUserImageView.widthAnchor.constraint(equalToConstant: 20)
-            ])
-            profileUserImageView.layer.cornerRadius = 10
-            self.clipsToBounds = true
-        }
-         let whoAndWhenPost = GLabel(font: UIFont.systemFont(ofSize: 12, weight: .light),
-                                            fontColor: .gGray)
-        
-         let likesNumber = GLabel(font: .regular14)
-        
-         let commentNumber = GLabel(font: .regular14)
-        
-         let commentImage = GLikeCommentButton(icon: UIImage(systemName: "heart"))
-         let heartImage = GLikeCommentButton(icon: UIImage(systemName: "text.bubble"))
-        
-        override init(frame: CGRect) {
-            super.init(frame: frame)
-            setup()
-        }
-        
-        required init(coder: NSCoder) {
-            fatalError("init(coder:) has not been implemented")
-        }
-        
-        private func setup() {
-            likesNumber.setSize(width: 30)
-            likesNumber.textAlignment = .left
-            commentNumber.setSize(width: 30)
-            commentNumber.textAlignment = .left
-            let stack = UIStackView(arrangedSubviews: [profileUserImageView,whoAndWhenPost,heartImage, likesNumber, commentImage,commentNumber])
-            stack.alignment = .fill
-            stack.distribution = .fillProportionally
-            stack.spacing = 5
-            stack.embed(in: self, with: .padding(top: 15, right: 10, bottom: 10, left: 10))
-        }
-        
-        func setupContent(whoAndWhenPost: String,
-                          profileUserImageView: UIImage?,
-                          likes: Int,
-                          comments: Int) {
-            self.profileUserImageView.image = profileUserImageView
-            self.whoAndWhenPost.text = whoAndWhenPost
-            commentNumber.text = String(comments)
-            likesNumber.text = String(likes)
-        }
-    }
 }
 
 //MARK: - TestCollectionView
@@ -215,7 +112,7 @@ class TestCollectionViewII: UIViewController, UICollectionViewDataSource, UIColl
         layout.sectionInset = UIEdgeInsets(top: 0, left: 0, bottom: 0 , right: 0)
         layout.itemSize = CGSize(width: 350, height: 200)
         
-       
+
         
         collectionView = UICollectionView(frame: self.view.frame, collectionViewLayout: layout)
         collectionView.dataSource = self
@@ -236,12 +133,9 @@ extension TestCollectionViewII {
     
     func collectionView(_ collectionView: UICollectionView, cellForItemAt indexPath: IndexPath) -> UICollectionViewCell {
         let cell = collectionView.dequeueReusableCell(withReuseIdentifier: cellIdentifier, for: indexPath as IndexPath) as! PostCollectionViewCell
-        cell.setupContentForSearch(image: UIImage(named: "sight"),
-                          name: "Московский Кремль",
-                          descriptionOfPlace: "Моско́вский Кремль — крепость в центре Москвы и древнейшая её часть, главный общественно-политический и историко-художественный комплекс города, официальная резиденция Президента Российской Федерации, вплоть до распада СССР в декабре 1991 года была официальной резиденцией Генерального секретаря ЦК КПСС.",
-                          whoAndWhenPost: "Artyom Korotkov",
-                          profileUserImageView: UIImage(named: "sight"),
-                          likes: 8, comments: 8)
+        cell.setupContent(image: UIImage(named: "sight"),
+                                   name: "Московский Кремль",
+                                   descriptionOfPlace: "Моско́вский Кремль — крепость в центре Москвы и древнейшая её часть, главный общественно-политический и историко-художественный комплекс города, официальная резиденция Президента Российской Федерации, вплоть до распада СССР в декабре 1991 года была официальной резиденцией Генерального секретаря ЦК КПСС.")
         return cell
     }
 }
