+++ GuidancerUI/CollectionViewCell/SearchScreenCollectionViewCell.swift
@@ -8,11 +8,87 @@
 import UIKit
 
 class SearchScreenCollectionViewCell: PostCollectionViewCell {
-    
+
+    private let informationInPost = ChangableInfoOfPostSV()
+
     override func setup() {
-        let sv = UIStackView(arrangedSubviews: [informationInPost])
-       sv.axis = .vertical
-       sv.embed(in: contentView)
+        super.setup()
+        contentStackView.addArrangedSubview(informationInPost)
     }
+
+    func setupContent(image: UIImage?,
+                      name: String,
+                      descriptionOfPlace: String,
+                      whoAndWhenPost: String,
+                      profileUserImageView: UIImage?,
+                      likes: Int,
+                      comments: Int
+    ) {
+        super.setupContent(image: image,
+                           name: name,
+                           descriptionOfPlace: descriptionOfPlace)
+        informationInPost.setupContent(whoAndWhenPost: whoAndWhenPost,
+                                       profileUserImageView: profileUserImageView,
+                                       likes: likes, comments: comments)
+    }
+
+    private final class ChangableInfoOfPostSV: UIView {
+
+         var profileUserImageView: UIImageView = {
+            let image = UIImageView()
+            image.frame.size = CGSize(width: 50, height: 50)
+             image.clipsToBounds = true
+            return image
+        }()
+
+        let whoAndWhenPost = GLabel(font: UIFont.systemFont(ofSize: 12, weight: .light),
+                                    fontColor: .gGray)
+        let likesNumber = GLabel(font: .regular14)
+        let commentNumber = GLabel(font: .regular14)
+        let commentImage = GLikeCommentButton(icon: UIImage(systemName: "heart"))
+        let heartImage = GLikeCommentButton(icon: UIImage(systemName: "text.bubble"))
+
+        override init(frame: CGRect) {
+            super.init(frame: frame)
+            setup()
+        }
+
+        required init(coder: NSCoder) {
+            fatalError("init(coder:) has not been implemented")
+        }
+
+        override func layoutSubviews() {
+            super.layoutSubviews()
+            NSLayoutConstraint.activate([
+                profileUserImageView.heightAnchor.constraint(equalToConstant: 50),
+                profileUserImageView.widthAnchor.constraint(equalToConstant: 20)
+            ])
+            profileUserImageView.layer.cornerRadius = 10
+            self.clipsToBounds = true
+        }
+
+        private func setup() {
+            likesNumber.setSize(width: 30)
+            likesNumber.textAlignment = .left
+            commentNumber.setSize(width: 30)
+            commentNumber.textAlignment = .left
+            let stack = UIStackView(arrangedSubviews: [profileUserImageView,whoAndWhenPost,heartImage, likesNumber, commentImage,commentNumber])
+            stack.alignment = .fill
+            stack.distribution = .fillProportionally
+            stack.spacing = 5
+            stack.embed(in: self, with: .padding(top: 15, right: 10, bottom: 10, left: 10))
+        }
+
+        func setupContent(whoAndWhenPost: String,
+                          profileUserImageView: UIImage?,
+                          likes: Int,
+                          comments: Int) {
+            self.profileUserImageView.image = profileUserImageView
+            self.whoAndWhenPost.text = whoAndWhenPost
+            commentNumber.text = String(comments)
+            likesNumber.text = String(likes)
+        }
+    }
+
 }
